--- devbox_client/src/assets/css/contact.css
@@ -1,54 +1,0 @@
-
-.contact-background {
-  position: relative;
-  height: 300px; /* 원래 높이로 복원 */
-  width: 100%;
-  overflow: hidden;
-}
-
-.contact-background img {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  z-index: 1;
-}
-
-.contact-background::before {
-  content: "";
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 50%; /* 원래대로 좌측 절반만 그라데이션 적용 */
-  height: 100%;
-  background: linear-gradient(to right, rgba(0, 0, 0, 0) 0%, rgba(0, 0, 0, 0.9) 100%);
-  z-index: 0;
-}
-
-.contact-header {
-  position: relative;
-  z-index: 2;
-  color: #fff;
-  font-family: 'PTBandocheB'; /* 폰트 적용 */
-}
-
-.bg-light {
-  background-color: #f8f9fa;
-}
-
-.text-primary {
-  font-family: 'PTBandocheB'; /* 폰트 적용 */
-  color: #007bff;
-}
-
-.text-muted {
-  color: #6c757d;
-}
-
-.light-300 {
-  font-weight: 300;
-}
-
-/* 추가 스타일을 여기에 작성하세요 */
