+++ devbox_client/src/components/Header.jsx
@@ -103,7 +103,7 @@ const Header = () => {
           <div className="navbar align-self-center d-flex">
             {user ? (
               <>
-                {/* <MsgBell /> */}
+                <MsgBell />
                 <a className="nav-link" href="/mypage">
                   <i className="bx bx-user-circle bx-sm text-primary"></i>
                 </a>
