+++ devbox_client/src/pages/message/MsgBell.jsx
@@ -1,38 +1,48 @@
-import { useEffect, useState } from "react";
-import { useLocation, useNavigate } from "react-router-dom";
+import React, { useEffect, useState } from "react";
+import { useUser } from '../../components/context/UserContext'
 
 const MsgBell = () => {
-    const navigate = useNavigate();
     const [nullReadTimeCount, setNullReadTimeCount] = useState(0);
-    // const { user } = useUser();
-    // const [bellData, setBellData] = useState({});
-
-    async function get() {
-        // if (user) {
-        try {
-            const res = await fetch('http://localhost:8080/msg/bell?reciver=민준');
-            const data = await res.json();
-            // setBellData(data);
-            if (Array.isArray(data.messages)) {
-                const nullCount = data.messages.filter(
-                    item => item.readTime === null
-                ).length;
-                setNullReadTimeCount(nullCount);
-            } else {
-                console.error("Received data is not in expected format:", data);
-            }
-        } catch (error) {
-            console.error("Failed to fetch data:", error);
-        }
-    }
+    const { user } = useUser();
 
     useEffect(() => {
-        get();
-    }, []);
+        const getMessages = async () => {
+            if (!user) return; // 사용자가 로그인하지 않았을 경우 fetch를 수행하지 않음
+
+            try {
+                const res = await fetch(`http://localhost:8080/msg/bell?reciver=${user.username}`, {
+                    method: 'GET',
+                    credentials: 'include',
+                    headers: {
+                        Authorization: `Bearer ${localStorage.getItem("accessToken")}`,
+                    },
+                });
+
+                if (!res.ok) {
+                    throw new Error('메시지 가져오기 실패');
+                }
+
+                const data = await res.json();
+
+                if (Array.isArray(data.messages)) {
+                    const unreadCount = data.messages.filter(
+                        (item) => item.readTime === null
+                    ).length;
+                    setNullReadTimeCount(unreadCount);
+                } else {
+                    console.error("예상하지 못한 데이터 형식입니다:", data);
+                }
+            } catch (error) {
+                console.error("데이터를 가져오는데 실패했습니다:", error);
+            }
+        };
+
+        getMessages();
+    }, [user]);
 
     return (
         <a className="nav-link" href="/message/list">
-            <i className='bx bx-bell bx-sm bx-tada-hover text-primary' style={{ transform: 'translate(0%, 0%)'}}>
+            <i className='bx bx-bell bx-sm bx-tada-hover text-primary' style={{ transform: 'translate(0%, 0%)' }}>
                 {nullReadTimeCount > 0 && (
                     <span className="badge rounded-pill bg-danger" style={{ position: 'absolute', top: '-10px', right: '-15px', fontSize: '12px' }}>
                         {nullReadTimeCount}
